1.load the older status timeline 2.add logging interceptor to retrofit 3.span can change it's background when get touched 4.add bi-directional SwipeRefreshLayout module 5.load the status pictures in middle size 6.display the retweeted status 7.display the hot rank of status 8.display the verified icon for the user;

+++ app/src/main/java/com/gnayils/obiew/bean/Timeline.java
@@ -11,7 +11,7 @@ import java.util.List;
 public class Timeline {
 
     public boolean hasvisible;
-    public int previous_cursor;
+    public long previous_cursor;
     public long next_cursor;
     public int total_number;
     public int interval;
